Adding support for classes
