+++ Grenat.Functional.DDD.Tests/Usings.cs
@@ -1,4 +1,5 @@
 global using System.Collections.Immutable;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using static Grenat.Functional.DDD.Option;
-global using static Grenat.Functional.DDD.Entity;
+global using static Grenat.Functional.DDD.Src.EntityOperations;
+global using static Grenat.Functional.DDD.Src.EntitySetters;
