+++ SampleProject/Usings.cs
@@ -2,6 +2,8 @@
 
 global using Grenat.Functional.DDD;
 global using static Grenat.Functional.DDD.Option;
+global using static Grenat.Functional.DDD.Src.EntityOperations;
+global using static Grenat.Functional.DDD.Src.EntitySetters;
 
 global using SampleProject.Domain.Entities;
 global using SampleProject.Domain.ValueObjects;
